add probability dist logic and test txt file

+++ encode.h
@@ -6,8 +6,9 @@
 #include <fstream>
 #include <string>
 #include <map>
+#include <iostream>
 
 bool encodeFile(std::string filename){};
-std::map<char, int> retrieveCharDistribution(std::vector<std::vector<char>> content){};
+std::map<char, float> retrieveCharDistribution(std::vector<std::vector<char>> content){};
 
 #endif
